+++ uml_router/uml_switch.c
@@ -7,6 +7,7 @@
 #include <stdlib.h>
 #include <signal.h>
 #include <fcntl.h>
+#include <stdint.h>
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <sys/poll.h>
@@ -41,7 +42,7 @@ struct request_v0 {
 #define SWITCH_MAGIC 0xfeedface
 
 struct request_v1 {
-  unsigned long magic;
+  uint32_t magic;
   enum request_type type;
   union {
     struct {
@@ -52,8 +53,8 @@ struct request_v1 {
 };
 
 struct request_v2 {
-  unsigned long magic;
-  int version;
+  uint32_t magic;
+  uint32_t version;
   enum request_type type;
   struct sockaddr_un sock;
 };
@@ -64,8 +65,8 @@ struct reply_v2 {
 };
 
 struct request_v3 {
-  unsigned long magic;
-  int version;
+  uint32_t magic;
+  uint32_t version;
   enum request_type type;
   struct sockaddr_un sock;
 };
